requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
deepgram-sdk==3.2.7
openai==1.3.7
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.0
asyncio==3.4.3 